Labs2025

Description

Módulo com os tipos de dados que vão ser utilizados para modelar as tarefas do trabalho prático de LI1/LP1 em 2025/26.

Documentation

data Terreno

Um tipo de terreno do mapa.

Constructors

Ar
Terreno vazio.

Agua
Terreno que afoga minhocas.

Terra
Terreno opaco e destrutivel.

Pedra
Terreno opaco e indestrutivel.

data TipoArma

Os diversos tipos de arma disponíveis para uma minhoca.

Constructors

Jetpack

Escavadora

Bazuca

Mina

Dinamite

data Estado

Estado do jogo.

Constructors

Estado

Fields

mapaEstado :: Mapa

O mapa atual.

objetosEstado :: [Objeto]

Uma lista com os objetos presentes no mapa. Para as funções que vai desenvolver, deve considerar que a ordem dos elementos é irrelevante.

minhocasEstado :: [Minhoca]

Uma lista com as minhocas no jogo. A ordem dos elementos é relevante, no sentido cada minhoca vai ser identificada pelo seu índice na lista.

data Jogada

Uma jogada que uma minhoca pode efetuar.

Constructors

Dispara TipoArma Direcao
Disparar uma arma numa dada direção.

Move Direcao
Mover-se numa dada direção.

data Minhoca

O estado completo de uma minhoca.

Constructors

Minhoca

Fields

posicaoMinhoca :: Maybe Posicao

Uma posição no mapa. Opcional porque a minhoca pode ter saído do mapa.

vidaMinhoca :: VidaMinhoca

O estado de saúde da minhoca.

jetpackMinhoca :: Int

Munições de Jetpack.

escavadoraMinhoca :: Int

Munições de Escavadora.

bazucaMinhoca :: Int

Munições de Bazuca.

minaMinhoca :: Int

Munições de Mina.

dinamiteMinhoca :: Int

Munições de Dinamite.

data Objeto

Um objeto colocado no mapa.

Constructors

Disparo
Um disparo de uma arma.

Fields

posicaoDisparo :: Posicao

A posição do disparo no mapa.

direcaoDisparo :: Direcao

A direção do disparo.

tipoDisparo :: TipoArma

O tipo de arma do disparo.

tempoDisparo :: Maybe Ticks

O tempo até o disparo explodir. Opcional porque nem todos os disparos de todas as armas têm um tempo pré-definido para explodir.

donoDisparo :: NumMinhoca

A minhoca que efetuou o disparo.

Barril
Um barril de pólvora.

Fields

posicaoBarril :: Posicao

A posição do barril no mapa.

explodeBarril :: Bool

Se o barril está prestes a explodir ou não.

data VidaMinhoca

O estado de saúde de uma minhoca.

Constructors

Viva Int
Está viva com um número inteiros de pontos de vida.

Morta
Está morta.

type Mapa = Matriz Terreno

O mapa do jogo é uma matriz de terrenos.

type NumMinhoca = Int

O índice de uma minhoca na lista de minhocas.

type NumObjeto = Int

O índice de um objeto na lista de objetos.

type Ticks = Int

Um tick é a unidade de tempo do jogo.

module Tarefa0_geral