Main

Description

Este módulo contém a função principal do jogo.

Responsabilidades principais:

Inicialização do estado inicial do jogo

Carregamento dos assets gráficos

Arranque do ciclo principal do Gloss

Ligação entre lógica, desenho e eventos

Este módulo não contém lógica de jogo, funcionando apenas como orquestrador dos vários subsistemas.

Configurações Base

window :: Display

Configuração da janela do jogo

background :: Color

Cor de fundo

fps :: Int

Taxa de atualização

Função principal

main :: IO ()

Carregar imagens

carregarAssets :: IO Assets

carregarImagem :: FilePath -> String -> IO (Maybe Picture)

Carrega uma imagem PNG do disco

carregarImagemOpcional :: FilePath -> String -> IO (Maybe Picture)

Carrega imagem opcional sem mostrar erro

Desenha o estado atual do jogo

desenhar :: Assets -> EstadoJogo -> Picture

EVENTOS COM CONTADOR

eventoComContador :: Assets -> IORef Int -> Event -> EstadoJogo -> IO EstadoJogo

eventoSelecaoContador :: IORef Int -> Event -> EstadoSelecao -> IO EstadoJogo

EVENTOS NA SELEÇÃO COM CONTADOR

eventoSelecaoMapa :: Event -> EstadoSelecaoMapa -> EstadoJogo

EVENTOS NA SELEÇÃO DE MAPA (TREINO) ^ mapaSelecionado: 0-4 = mapas, 5 = voltar

iniciarTreinoComMapa :: Int -> EstadoJogo

Inicia treino com mapa específico

eventoVictoryContador :: IORef Int -> Event -> EstadoFinal -> IO EstadoJogo

EVENTOS DE VITÓRIA COM CONTADOR

FUNÇÃO AUXILIAR: Inicia jogo com mapa específico

iniciarComMapa :: ModoJogo -> MapaCompleto -> Int -> EstadoJogo

Atualiza o estado do jogo ao longo do tempo

atualizar :: Assets -> Float -> EstadoJogo -> EstadoJogo